Track 1: Favor Bind Variables Over String Concatenation
It's a myth that SQL Injection is not possible in Java. This myth is
perpetuated by the fact that many Java developers typically interact
with the database through an ORM and thus aren't writing as many raw
SQL statements.

However, with tools like JDBC Template, the pendulum is swinging back
the other way. As more engineers elect to write out their SQL queries
the potential for SQL injection in Java increases.

The simple answer is bind variables in where clauses, which this
commit demonstrates. Regardless of the underlying language, the need
for bind variables is understood and should be favored, whether it is
through a third-party library or through core language offerings.

As a side note, this doesn't rid us of all potential problems with
concatenation since not everything can use a bind variable. Check out
the next commit for another option.