Track 2: Encode Outputs
Bind Variables should be the approach whenever it is supported.

But, when it isn't, for example when the FROM clause or other areas
are dynamic, then whitelist-validate the input and encode the output.

One way to encode the output is to use the ESAPI encoder, though its
support is limited to only a handful of databases. In this example, we
pretend that this would be connecting to a MySQL database and so it
uses the MySQLCodec.

Check other branches to learn about validating input.