Track 3: Clarify Query Expectations
It is better to return an error than to accidentally return data that
belongs to one user to another.

To that end, while not directly related to SQL injection, it is more
secure to ensure that only one result is returned when that is all
that was intended. Indeed, because the code originally simply returned
the first result that came back, in an injection scenario this meant
the first user in the entire database.

Further, queries that simply select for * and use
ResultSet#getString(int) are vulnerable to potential
information leakage as well as an overall brittleness that can be
leveraged by intruders. A clearer approach is to specify the columns
in the SELECT clause and to use ResultSet#getString(String).