Track 5: Disable Literals at the DB Level
SQL Injection can also be mitigated at the DB level in some cases.

For example, H2 ships with a feature that allows a user to disallow
any literals in a query. This means that a query like
select * from users where id = '1' would fail. The database would only
accept a query like select * from users where id = ? with the value
being passed as a parameter.

This only mitigates SQL injection where bind variables would resolve
the issue, but is a nice additional defense.