Track 0: Setup
A request is vulnerable to enumeration when there is some aspect of
its response based differs based on whether or not the input data is
valid, revealing otherwise secret information.

In the case of login, this is where we can find out all the valid
usernames simply by putting in random usernames and passwords and
seeing if the error message differs. The same holds true for the
forgot password page as well as other pages that very a request
against persistent state.

The Terracotta base is only vulnerable in the registration case; this
commit makes login vulnerable, too. Note, ironically, that it was
made vulnerable to enumeration actually by taking a step forward with
respect to password hashing.