Track 1: Unify Messaging
The first and often easiest way to combat Enumeration is to unify
messaging across error responses. This means making error messages
more general so that an intruder cannot infer secret information
through inferences about what the website did *not* say.

In these cases, it means the messaging to the user, but it can also
refer to status codes, like 404 vs 403, and headers.

This won't stop timing attacks, though, so check out the next commit
for more strategy.