Track 2: Use Constant-Time Algorithms
Given a set of bad-user failed request times and a set of bad-password
failed request times, with what confidence could you infer which
belong to which set?

The further apart these request times travel from one another, the
more it is just like the "username or password" problem.

To repair, we can sometimes offload the work onto a separate thread or
out to a separate service like Kafka. Or, we can sometimes perform the
same work on a random set of data as we do here in UserService.

It's also good practice to trim the database load where possible,
which is why UserService#findByUsernameAndPassword changes to query
only for the password and to return only a boolean.
wq