Step 0: Setup
An application is vulnerable to brute force authentication for the
same reasons that an adult is vulnerable to brute force demands from a
three-year old. Simply put, if a client is permitted to ask over and
over until it finds the weak spot, then it is just a matter of time.

Terracotta is vulnerable in numerous ways to authentication brute-
forcing via its four login endpoints. There is a vanilla script called
BruteForceTester that you are welcome to customize for your needs. Or
you can just do what my 14-year-old son did to break in without any
programming knowledge.

Make sure to check the rest of the lesson for tips and fixes.