Step 1: Replace In-Code Trojans
Applications often have backdoor steps that can be taken in order to
allow developers and operators into the system. And since backdoors
like these are often created either under duress or early on in the
project before it's really matured, these backdoors often have
terrible or non-existent password policies.

Even worse, it's not uncommon for these to appear somewhere in the
codebase.

While introducing a fully-fledged impersonation system is out-of-
scope for this lesson, we can at least introduce some kind of reason-
able password policy and also remove the passwords from the codebase.

This commit does this, understanding that ultimately the contents of
application.properties could be further secured by managing the
composition of the file through Chef, Salt, or Pepper.