Step 2: Force Update of Other Trojans
Some trojans are simply default usernames and passwords that are put
in place to simplify installation and setup.

Now, there are a number of ways to force an update of these kinds of
credentials including alerting on startup or introducing business
logic that forces the end user to make the change after logging in.

As a bit of a last resort, the code can also detect and make the
change itself, which is what this commit demonstrates. The code, if
the password is the default, will update the password to a random
one. It's not the best choice, but it is a step up from leaving it as
a default value.