Step 3: Soft Lockouts
Eliminating bad passwords is a good first step, but to really defend
against brute force, we need to introduce a way to throttle clients
when they are making too many requests.

In the case of authentication, this is often materialized as tracking
the number of consecutive failed attempts that a client has made. A
hard cap works well in many cases, though some enterprises attempt to
be more clever, using exponential backoff or additional login
requirements once brute-forcing is suspected (like requiring a captcha
in order to confirm that it isn't a bot).

To use a cap, though, we'd ideally like to remember the principle of
enumeration, too, though. If all we have to do is try to login five
times and then the application will inform us that the username
doesn't exist (by stating that the account has been locked), then that
is a simple vector for an attacker to confirm a valid username.

To that end, this soft lockout uses a cache to collect login attempts
regardless of whether they are against a valid account or not, meaning
that invalid accounts will get "locked out" as well.