Step 4: IP Soft Lockout
Of course, just detecting when a single username is being bruteforced
is not enough. We need to also detect cross-account attacks.

One way we can mitigate against this kind of attack is by tracking ip
addresses. Hopefully, we have detection like this at the network
level, but this commit is a representation of what something like that
could look like at the software level.

Play close attention to the implementation of IPAddressResolver, which
is cautious about trusting the X-Forwarded-For header.