Step 0: Demo Support Classes
The JCA classes are very noisy when it comes to checked exceptions.

For that reason, these demos use a wrapped version of several of the
JCA classes. These aren't intended for production use, and are
simply meant to remove some of the try-catch blocks that make the
core logic harder to read.