Step 1: Storing Keys in Java
The JCA has both keys and specs, with specs being a storage-
friendly representation.

To generate keys, use KeyGenerator for symmetric keys and
KeyPairGenerator for asymmetric keys. When generating, make sure to
indicate the size of key so that code isn't dependent on JVM
defaults.

To trade keys for specs and vice-versa, use SecretKeyFactory for
symmetric keys and KeyFactory for asymmetric keys. Though, note that
factories are typically not used for symmetric keys.