Step 2: Using the KeyStore API
The KeyStore API can be used to interact with a KeyStore
programmatically, as an alternative to the Command-Line interface.

KeyStores are limited as a general-purpose key storage mechanism
since they assume that any private key will be accompanied with a
certificate. This example isn't meant to suggest this as a design
option; instead, it highlights some of the naunaces of the KeyStore
API.