Step 3: Using Vault in Java
Vault is a service that can be used to manage keys. It provides
support for generating, rotating, and exporting keys as well as
encryption and signing so that keys never need to leave Vault.

This commit demonstrates some simple usage of Vault, including
support for key rotation.

Also, it demonstrates the considerations we need to take into account
when reading it keys from an external system, especially when keys
are formatted in PKCS#1, which Java does not support natively.