Add CRLF Injection Vulnerability
Note that this vulnerable code is lifted from a very old version of
Tomcat and relies on a classloading trick to work. As such, this
vulnerability can only be tested when Tomcat is run as an embedded
container. Down the road, it would be better to find a more reliable
way to demonstrate CRLF.