Step 3: Turn Features Off
A slightly less powerful approach than disabling DOCTYPEs or using a
no-op EntityResolver is disabling each risky feature individually.

This is riskier for the same reason that blacklisting is risky. Other
attacks through DOCTYPEs may surface in the future, and it is better
to simply disallow them wholesale, if you are able.