Step 4: Explicitly Shut Off Dangerous Features
Some dangerous features are turned off by default, like XInclude. It's
valuable to turn them off explicitly so that it's clear what features
our code is and is not using. In the event that a future release turns
that feature on by default, it will remain off in your application.