Step 1: Symmetric Encryption with AES
AES is a common standard for performing symmetric encryption. It is
performed on the block level, meaning decryption can be streamed, and
is well-understood in the security community at large.

Note that when using the Cipher class that progressive calls to update
are returning encrypted or decrypted material - don't let that confuse
you from how JCA treats update in other JCA classes.