Step 3: Using GCM for Authenticated Encryption
Stronger (and faster) than CBC is GCM. It's stronger because it gives
us integrity of the IV and any other associated data. Happily, it's
also faster since it can be parallelized.

The specification indicates that the IV should be 12-bytes in length
for ideal performance.

AES GCM is the preferred algorithm for symmetric encryption.