Step 4: Hybrid Encryption
Hybrid Encryption is favorable to asymmetric encryption since it uses
symmetric encryption for the body, making it faster for larger
messages.

Basically, a new secret key is created for each message. This is used
to encrypt the body. Then asymmetric encryption is used to encrypt
that secret key. Both the encrypted message and the encrypted key are
sent.