Step 0: Adding Separate Service for TLS Demos
In preparation for these demos, I've added a simple remote service
that the AccountService coordinates with for making deposits.

In order to demonstrate more JSSE features, you'll notice that it
implements its own binary protocol over TCP since several of these
security features aren't exposed over HttpsURLConnection.

There are three key stores as part of this initial commit, each to
demonstrate a different aspect of securing TLS. The first is
client.jks, which is part of terracotta-bank-servlet. It contains a
test CA, very similar to the one created in the authentication-tls
branch. The second is server.jks, which contains a key signed by that
CA and which has a SAN of "localhost". The third is evil.jks, which
contains a key signed by that CA and which has a SAN of "evil.com".