Step 3: Don't use All-Trusting TrustManagers
Sometimes developers will use an all-trusting TrustManager because
they feel like they are in a pinch or they don't understand what they
need to do to make the SSL handshake work.

This shouldn't be done, especially because it is hardcoded, making it
tricky to ensure that said code doesn't make it all the way to
production.

A better solution is to load up the keystore that has the trusted
certificate or that has the trusted CA that signed the certificate.

For self-signed certificates, typical of the dev cycle, we can simply
use the same keystore on both the client and the server.

Of course, we may not have control of the server or we may have many
more points of contact than just one client and one server. In this
case, it may be better to create a test CA, adding that to the
client's keystore, and then signing all test certs with that CA.

That is the scenario demonstrated here.