Step 4: Configure TLS Version and Cipher Suites
By default, Java SSL clients will still communicate over TLSv1.0 if
the server requests it. While a server may be properly configured,
it's possible for a man in the middle to intercept the communication
during the negotiation step, downgrading the communication
to something lower.

If you aren't willing to communicate over TLSv1.0 (or 1.1), then say
so directly in your client configuration.

The same principle applies to cipher suites. State the set of cipher
suites you are willing to communicate in. These days, only select
suites that are ephermeral (i.e. have the property of forward secrecy)
and use authenticated encryption, like GCM or CCM.