Step 5: Configure for Hostname Verification
Hostnames are part of the HTTPS application protocol. Since TLS is
a network protocol, it doesn't think about hostnames.

In other words, without additional configuration, JSSE won't check the
hostname in the server's certificate to make sure that it matches the
hostname that the client is contacting.

If you are using HttpsURLConnection, this is setup automatically.
However, it's common to bypass this class for increased configurability,
meaning we need to remember this extra step.