Step 6: On JDK 11, Use TLSv1.3
TLSv1.3 is faster and more secure. It mandates ephemeral key exchange
as well as authenticated encryption. It removes deprecated cipher
suites, and it secures what negotiation happens in the clear.

It's supported in JDK 11, so when trying this step in the demo, make
sure you have that version or higher on your machine.