Step 1: Use 'TLS' for SSLContext algorithm
Using "TLSv1.2" as the algorithm in SSLContext has confusing
semantics. It doesn't mean that the connection should use TLSv1.2.

Instead, it means that the connection should negotiate for TLSv1.2,
negotiating down if the client or server doesn't support that version.

The danger in these semantics is downgrade attacks, which is treated
in a later commit. Here, though, we simply want to say "TLS" to
indicate that the version isn't being controlled here. We'll control
it in another place where the semantics are clearer.