Step 2: Use 'NewSunX509' KeyManagers
The default algorithm for KeyManager always selects the first key in
the keystore, whether or not it is expired and whether or not it has
the right keyUsage values. This makes key management a bit of a pain.

"NewSunX509" is a little bit better in that it will check the
expiration date and the keyUsage values, sending down the first
matching one that it finds.